requirements.txt
pygame-ce==2.5.3
pylint==3.3.4